Refactor file server tests to be more pytest-y
